+++ packages/sit-onyx/.storybook/main.ts
@@ -10,7 +10,12 @@ const onyxLayers = readFileSync(
 
 const config: StorybookConfig = {
   stories: ["./pages/*.mdx", "../src/**/*.stories.ts"],
-  addons: ["@storybook/addon-essentials", "storybook-dark-mode", "@storybook/addon-a11y"],
+  addons: [
+    "@storybook/addon-essentials",
+    "storybook-dark-mode",
+    "@storybook/addon-a11y",
+    "storybook-addon-tag-badges",
+  ],
   staticDirs: ["./public"],
   framework: {
     name: "@storybook/vue3-vite",