+++ packages/sit-onyx/src/components/OnyxDataGrid/OnyxDataGrid.vue
@@ -4,25 +4,32 @@
   generic="
     TEntry extends DataGridEntry,
     TTypeRenderer extends TypeRenderMap<TEntry>,
+    TColumnGroup extends ColumnGroupConfig,
     TFeatureName extends symbol,
     TFeatures extends DataGridFeature<TEntry, TTypeRenderer, TFeatureName>[] | []
   "
 >
 import { ref, toRefs, watch, type Ref, type WatchHandle } from "vue";
 import { injectI18n } from "../../i18n";
-import { useDataGridFeatures, type DataGridFeature, type TypeRenderMap } from "./features";
+import type { TableColumnGroup } from "../OnyxTable/types";
+import {
+  useDataGridFeatures,
+  type ColumnGroupConfig,
+  type DataGridFeature,
+  type TypeRenderMap,
+} from "./features";
 import OnyxDataGridRenderer from "./OnyxDataGridRenderer/OnyxDataGridRenderer.vue";
 import type { DataGridRendererColumn, DataGridRendererRow } from "./OnyxDataGridRenderer/types";
 import type { DataGridEntry, DataGridMetadata, OnyxDataGridProps } from "./types";
 
 const props = withDefaults(
-  defineProps<OnyxDataGridProps<TEntry, TTypeRenderer, TFeatureName, TFeatures>>(),
+  defineProps<OnyxDataGridProps<TEntry, TColumnGroup, TTypeRenderer, TFeatureName, TFeatures>>(),
   {
     features: () => [] as TFeatures,
   },
 );
 
-const { t } = injectI18n();
+const i18n = injectI18n();
 
 defineSlots<{
   /**
@@ -36,8 +43,9 @@ defineSlots<{
 // Using Ref types to avoid `UnwrapRef` issues
 const renderColumns: Ref<DataGridRendererColumn<TEntry>[]> = ref([]);
 const renderRows: Ref<DataGridRendererRow<TEntry, DataGridMetadata>[]> = ref([]);
+const rendererColumnGroups: Ref<TableColumnGroup[] | undefined> = ref();
 
-const { columns, data, features } = toRefs(props);
+const { columns, data, features, columnGroups } = toRefs(props);
 
 /**
  * Function to be able to reset the watcher in case of the features being updated.
@@ -47,12 +55,16 @@ const createFeatureBuilderWatcher = ({
   createRendererColumns,
   createRendererRows,
   watchSources,
-}: ReturnType<typeof useDataGridFeatures<TEntry, TFeatureName, TTypeRenderer, TFeatures>>) => {
+  createRendererColumnGroups,
+}: ReturnType<
+  typeof useDataGridFeatures<TEntry, TFeatureName, TTypeRenderer, TColumnGroup, TFeatures>
+>) => {
   return watch(
-    [data, t, ...watchSources],
+    [data, columns, columnGroups, i18n.locale, i18n.t, ...watchSources],
     () => {
       renderColumns.value = createRendererColumns();
       renderRows.value = createRendererRows(data.value);
+      rendererColumnGroups.value = createRendererColumnGroups();
     },
     { immediate: true, deep: true },
   );
@@ -61,7 +73,11 @@ const createFeatureBuilderWatcher = ({
 watch(
   features,
   () => {
-    const featureBuilder = useDataGridFeatures(features.value, { t, columnConfig: columns });
+    const featureBuilder = useDataGridFeatures(features.value, {
+      i18n,
+      columnConfig: columns,
+      columnGroups,
+    });
     disposeWatcher?.();
     disposeWatcher = createFeatureBuilderWatcher(featureBuilder);
   },
@@ -70,7 +86,11 @@ watch(
 </script>
 
 <template>
-  <OnyxDataGridRenderer :columns="renderColumns" :rows="renderRows">
+  <OnyxDataGridRenderer
+    :column-groups="rendererColumnGroups"
+    :columns="renderColumns"
+    :rows="renderRows"
+  >
     <template #empty>
       <slot name="empty" />
     </template>