+++ packages/sit-onyx/src/components/OnyxDataGrid/features/all.ts
@@ -2,6 +2,7 @@ export * from "./selection/types";
 export * from "./sorting/types";
 export * from "./stickyColumns/types";
 
+export { useFiltering } from "./filtering/filtering";
 export { useSelection } from "./selection/selection";
 export { useSorting } from "./sorting/sorting";
 export { useStickyColumns } from "./stickyColumns/stickyColumns";