+++ packages/sit-onyx/src/components/OnyxTable/OnyxTable.vue
@@ -1,10 +1,8 @@
 <script lang="ts" setup>
-import graphSearch from "@sit-onyx/icons/graph-search.svg?raw";
 import { computed, type VNode } from "vue";
 import { useDensity } from "../../composables/density";
 import { injectI18n } from "../../i18n";
 import OnyxEmpty from "../OnyxEmpty/OnyxEmpty.vue";
-import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
 import type { OnyxTableProps } from "./types";
 
 const props = withDefaults(defineProps<OnyxTableProps>(), {
@@ -82,12 +80,7 @@ const isEmptyMessage = computed(() => t.value("table.empty"));
               <td colspan="100%">
                 <div class="onyx-table__empty-content">
                   <slot name="empty" :default-message="isEmptyMessage">
-                    <OnyxEmpty>
-                      <template #icon>
-                        <OnyxIcon :icon="graphSearch" size="48px" />
-                      </template>
-                      {{ isEmptyMessage }}
-                    </OnyxEmpty>
+                    <OnyxEmpty> {{ isEmptyMessage }} </OnyxEmpty>
                   </slot>
                 </div>
               </td>
@@ -291,21 +284,19 @@ $border: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral)
 
     // column hover styles
     th:not(&__colgroup):hover::before {
-      background-color: color-mix(in srgb, var(--onyx-color-base-neutral-500), transparent 85%);
-      content: "";
-      height: 100vh;
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      bottom: 0;
-      // needed in order for other components like buttons etc. to be clickable and to prevent showing the column hover effect when hovering down over a row
-      pointer-events: none;
-    }
-
-    // hover styles are disabled when the table is empty.
-    &:has(&__empty) th {
-      pointer-events: none;
+      .onyx-table:not(:has(.onyx-table__empty)) & {
+        background-color: color-mix(in srgb, var(--onyx-color-base-neutral-500), transparent 85%);
+        content: "";
+        height: 100vh;
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        bottom: 0;
+        z-index: -1;
+        // needed in order for other components like buttons etc. to be clickable and to prevent showing the column hover effect when hovering down over a row
+        pointer-events: none;
+      }
     }
 
     &__colgroup {